requirements.txt
torch
torchvision
transformers
open_clip_torch
pytorch-tabnet
lifelines
shap
umap-learn
groq
streamlit
fastapi
uvicorn
python-dotenv
kagglehub
requests
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
lightgbm
plotly
py3Dmol
Pillow
tqdm
scipy
